# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "GraphEdgeDetailResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "contract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract"
          },
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "edge_type": {
            "title": "Edge Type",
            "type": "string"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "source_node_id": {
            "title": "Source Node Id",
            "type": "string"
          },
          "target_node_id": {
            "title": "Target Node Id",
            "type": "string"
          },
          "validation_record": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Record"
          },
          "validation_state": {
            "title": "Validation State",
            "type": "string"
          }
        },
        "required": [
          "edge_id",
          "source_node_id",
          "target_node_id",
          "edge_type",
          "origin",
          "validation_state",
          "confidence",
          "protocol",
          "contract",
          "evidence",
          "validation_record",
          "metadata"
        ],
        "title": "GraphEdgeDetailResponse",
        "type": "object"
      },
      "GraphEdgeProjectionResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "edge_type": {
            "title": "Edge Type",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "source_node_id": {
            "title": "Source Node Id",
            "type": "string"
          },
          "target_node_id": {
            "title": "Target Node Id",
            "type": "string"
          },
          "validation_state": {
            "title": "Validation State",
            "type": "string"
          }
        },
        "required": [
          "edge_id",
          "source_node_id",
          "target_node_id",
          "edge_type",
          "origin",
          "validation_state",
          "confidence",
          "protocol"
        ],
        "title": "GraphEdgeProjectionResponse",
        "type": "object"
      },
      "GraphHierarchyEdgeResponse": {
        "properties": {
          "child_node_id": {
            "title": "Child Node Id",
            "type": "string"
          },
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "parent_node_id": {
            "title": "Parent Node Id",
            "type": "string"
          }
        },
        "required": [
          "edge_id",
          "parent_node_id",
          "child_node_id"
        ],
        "title": "GraphHierarchyEdgeResponse",
        "type": "object"
      },
      "GraphNodeDetailResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "node_type": {
            "title": "Node Type",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "stable_key": {
            "title": "Stable Key",
            "type": "string"
          },
          "validation_record": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Record"
          },
          "validation_state": {
            "title": "Validation State",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "stable_key",
          "node_type",
          "name",
          "origin",
          "validation_state",
          "confidence",
          "evidence",
          "validation_record",
          "metadata"
        ],
        "title": "GraphNodeDetailResponse",
        "type": "object"
      },
...
      "GraphOverrideCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "created_by": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "operation": {
            "$ref": "#/components/schemas/OverrideOperation"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "operation"
        ],
        "title": "GraphOverrideCreateRequest",
        "type": "object"
      },
      "GraphOverrideResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "entity_kind": {
            "title": "Entity Kind",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "override_id": {
            "title": "Override Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "target_edge_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Edge Id"
          },
          "target_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Node Id"
          }
        },
        "required": [
          "override_id",
          "revision_id",
          "operation",
          "entity_kind",
          "target_node_id",
          "target_edge_id",
          "payload",
          "reason",
          "created_by",
          "created_at"
        ],
        "title": "GraphOverrideResponse",
        "type": "object"
      },
      "GraphProjectionResponse": {
        "properties": {
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeProjectionResponse"
            },
            "title": "Edges",
            "type": "array"
          },
          "graph_id": {
            "title": "Graph Id",
            "type": "string"
          },
          "hierarchy": {
            "items": {
              "$ref": "#/components/schemas/GraphHierarchyEdgeResponse"
            },
            "title": "Hierarchy",
            "type": "array"
          },
          "issue_counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Issue Counters",
            "type": "object"
          },
          "layout_hints": {
            "additionalProperties": true,
            "title": "Layout Hints",
            "type": "object"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeProjectionResponse"
            },
            "title": "Nodes",
            "type": "array"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "repository_state_id": {
            "title": "Repository State Id",
            "type": "string"
          },
          "revision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "graph_id",
          "audit_id",
          "project_id",
          "repository_state_id",
          "schema_version",
          "status",
          "summary",
          "nodes",
          "edges",
          "hierarchy"
        ],
        "title": "GraphProjectionResponse",
        "type": "object"
      },
      "GraphRevisionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "created_by": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "parent_revision_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision Id"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "GraphRevisionCreateRequest",
        "type": "object"
      },
      "GraphRevisionResponse": {
        "properties": {
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "graph_id": {
            "title": "Graph Id",
            "type": "string"
          },
          "parent_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision Id"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "revision_id",
          "graph_id",
          "parent_revision_id",
          "status",
          "title",
          "created_by",
          "created_at",
          "committed_at"
        ],
        "title": "GraphRevisionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OverrideOperation": {
        "enum": [
          "ADD_NODE",
          "UPDATE_NODE",
          "MOVE_NODE",
          "SUPPRESS_NODE",
          "RESTORE_NODE",
          "ADD_EDGE",
          "UPDATE_EDGE",
          "SUPPRESS_EDGE",
          "RESTORE_EDGE"
        ],
        "title": "OverrideOperation",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Architecture Visualizer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audits/{audit_id}/graph": {
      "get": {
        "operationId": "get_audit_graph_api_v1_audits__audit_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProjectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit Graph",
        "tags": [
          "graphs"
        ]
      }
    },
...
    "/api/v1/graphs/{graph_id}/edges/{edge_id}": {
      "get": {
        "operationId": "get_graph_edge_api_v1_graphs__graph_id__edges__edge_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "title": "Graph Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "title": "Edge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphEdgeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph Edge",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/v1/graphs/{graph_id}/nodes/{node_id}": {
      "get": {
        "operationId": "get_graph_node_api_v1_graphs__graph_id__nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "title": "Graph Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphNodeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph Node",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/v1/graphs/{graph_id}/revisions": {
      "get": {
        "operationId": "list_graph_revisions_api_v1_graphs__graph_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "title": "Graph Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GraphRevisionResponse"
                  },
                  "title": "Response List Graph Revisions Api V1 Graphs  Graph Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Graph Revisions",
        "tags": [
          "graphs"
        ]
      },
      "post": {
        "operationId": "create_graph_revision_api_v1_graphs__graph_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "title": "Graph Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphRevisionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Graph Revision",
        "tags": [
          "graphs"
        ]
      }
    },
...
    "/api/v1/revisions/{revision_id}/commit": {
      "post": {
        "operationId": "commit_graph_revision_api_v1_revisions__revision_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Graph Revision",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/v1/revisions/{revision_id}/manual-nodes/{node_id}": {
      "delete": {
        "operationId": "delete_unsaved_manual_node_api_v1_revisions__revision_id__manual_nodes__node_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Unsaved Manual Node",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/v1/revisions/{revision_id}/overrides": {
      "post": {
        "operationId": "create_graph_override_api_v1_revisions__revision_id__overrides_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphOverrideCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphOverrideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Graph Override",
        "tags": [
          "graphs"
        ]
      }
    }
  }
}